- Message lookups are now cached, so safe when IRQs threaded.
  - AddCallback in finalise corrected to RemoveCallback
  - RAM rule added to makefile for making ROM replacement

Detail:
  Customer I reported that the STB-3 was suffering from IRQ reentrancy problems.
Some of these were tracked down to PortManager calling MessageTrans_LookUp
(when IRQs were threaded), which can cause file operations after
Service_ResourceFSStarting has been issued, invalidating the MessageTrans FD.

  PortManager now caches the contents of its Tags file upon startup in a
linked list, which it searches instead of using MessageTrans.

  During testing, it was also found that crashes would often occur when
killing the module.  This was tracked down to an erroneous OS_AddCallback in
the finalisation handler.

Admin:
  Tested on STB-3.  Rentrancy problems during booting fixed.  The module
should be loaded _very_ early on in the boot sequence.  Preferrably a the
module should be launched from a utility which peeks the current softcopy of
the port bits, and programs the module with these as it's initial value for
the port.


Version 0.15. Tagged as 'PortMan-0_15'